Move CommitishComparator and GitChangeSetFilter to where they are needed
