+++ package.json
@@ -17,16 +17,16 @@
     "eslint-config-crockford": "2.0.0",
     "eslint-config-standard": "14.1.1",
     "eslint-plugin-import": "2.22.0",
-    "eslint-plugin-jest": "24.1.3",
+    "eslint-plugin-jest": "24.7.0",
     "eslint-plugin-jsdoc": "30.7.8",
     "eslint-plugin-node": "11.1.0",
     "eslint-plugin-promise": "4.2.1",
     "eslint-plugin-standard": "4.0.1",
     "jest": "26.6.3",
-    "stylelint": "13.9.0",
+    "stylelint": "13.13.1",
     "stylelint-config-recommended": "3.0.0",
     "stylelint-config-standard": "20.0.0",
-    "stylelint-scss": "3.19.0",
+    "stylelint-scss": "3.21.0",
     "typescript": "3.9.7"
   },
   "devDependencies": {}
