+++ src/main/php/CommandLine/FileFinders/AllCheckableFileFinder.php
@@ -6,7 +6,6 @@ namespace Zooroyal\CodingStandard\CommandLine\FileFinders;
 
 use Zooroyal\CodingStandard\CommandLine\Git\GitChangeSet;
 use Zooroyal\CodingStandard\CommandLine\Git\GitChangeSetFactory;
-use Zooroyal\CodingStandard\CommandLine\Git\GitChangeSetFilter;
 use Zooroyal\CodingStandard\CommandLine\Process\ProcessRunner;
 
 class AllCheckableFileFinder implements FileFinderInterface
