+++ tests/Unit/CommandLine/FileFinders/AdaptableFileFinderTest.php
@@ -10,8 +10,8 @@ use PHPUnit\Framework\TestCase;
 use Symfony\Component\Console\Exception\InvalidArgumentException;
 use Zooroyal\CodingStandard\CommandLine\FileFinders\AdaptableFileFinder;
 use Zooroyal\CodingStandard\CommandLine\FileFinders\AllCheckableFileFinder;
+use Zooroyal\CodingStandard\CommandLine\FileFinders\CommitishComparator;
 use Zooroyal\CodingStandard\CommandLine\FileFinders\DiffCheckableFileFinder;
-use Zooroyal\CodingStandard\CommandLine\Git\CommitishComparator;
 use Zooroyal\CodingStandard\CommandLine\Git\GitChangeSet;
 use Zooroyal\CodingStandard\CommandLine\Git\GitInputValidator;
 use Zooroyal\CodingStandard\Tests\Tools\SubjectFactory;
