+++ tests/Unit/CommandLine/FileFinders/AllCheckableFileFinderTest.php
@@ -8,9 +8,9 @@ use Mockery;
 use Mockery\MockInterface;
 use PHPUnit\Framework\TestCase;
 use Zooroyal\CodingStandard\CommandLine\FileFinders\AllCheckableFileFinder;
+use Zooroyal\CodingStandard\CommandLine\FileFinders\GitChangeSetFilter;
 use Zooroyal\CodingStandard\CommandLine\Git\GitChangeSet;
 use Zooroyal\CodingStandard\CommandLine\Git\GitChangeSetFactory;
-use Zooroyal\CodingStandard\CommandLine\Git\GitChangeSetFilter;
 use Zooroyal\CodingStandard\CommandLine\Process\ProcessRunner;
 use Zooroyal\CodingStandard\Tests\Tools\SubjectFactory;
