+++ tests/Unit/CommandLine/FileFinders/CommitishComparatorTest.php
@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
-namespace Zooroyal\CodingStandard\Tests\Unit\CommandLine\Git;
+namespace Zooroyal\CodingStandard\Tests\Unit\CommandLine\FileFinders;
 
 use Mockery;
 use Mockery\MockInterface;
 use PHPUnit\Framework\TestCase;
-use Zooroyal\CodingStandard\CommandLine\Git\CommitishComparator;
+use Zooroyal\CodingStandard\CommandLine\FileFinders\CommitishComparator;
 use Zooroyal\CodingStandard\CommandLine\Git\GitInputValidator;
 use Zooroyal\CodingStandard\CommandLine\Process\ProcessRunner;
 use Zooroyal\CodingStandard\Tests\Tools\SubjectFactory;
@@ -21,7 +21,8 @@ class CommitishComparatorTest extends TestCase
     protected function setUp(): void
     {
         $subjectFactory = new SubjectFactory();
-        $buildFragments = $subjectFactory->buildSubject(CommitishComparator::class);
+        $buildFragments = $subjectFactory->buildSubject(
+            \Zooroyal\CodingStandard\CommandLine\FileFinders\CommitishComparator::class);
         $this->subject = $buildFragments['subject'];
         $this->subjectParameters = $buildFragments['parameters'];
     }
